Renamed ShipTests to ShipTestsDepricated and ECSShipTests to ShipTests. Modified ShipFactory to use Entity instead of ID or guid. Added few constructors for copy to new datablobs.
