+++ Pulsar4X/Pulsar4X.ECSLib/DataBlobs/MassVolumeDB.cs
@@ -55,5 +55,11 @@ namespace Pulsar4X.ECSLib
             Mass = mass;
             Density = density;
         }
+
+        public MassVolumeDB(MassVolumeDB massVolumeDB)
+            :this(massVolumeDB.Mass, massVolumeDB.Density)
+        {
+            
+        }
     }
 }
