+++ Pulsar4X/Pulsar4X.ECSLib/DataBlobs/TreeHierarchyDB.cs
@@ -130,5 +130,10 @@ namespace Pulsar4X.ECSLib
         {
 
         }
+
+        public ConcreteTreeHierarchyDB(ConcreteTreeHierarchyDB concreteTreeHierarchyDB)
+            : base(concreteTreeHierarchyDB.Parent)
+        {
+        }
     }
 }
