+++ Pulsar4X/Pulsar4X.Tests/Pulsar4X.Tests.csproj
@@ -53,13 +53,13 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="DataBlobTests.cs" />
-    <Compile Include="ECSShipTests.cs" />
+    <Compile Include="ShipTests.cs" />
     <Compile Include="EngineCommsTests.cs" />
     <Compile Include="EntityManagerTests.cs" />
     <Compile Include="MainGameLoopTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SaveGameTests.cs" />
-    <Compile Include="ShipTests.cs" />
+    <Compile Include="ShipTestsDepricated.cs" />
     <Compile Include="StorageTests.cs" />
     <Compile Include="UnitHelperTests.cs" />
     <Compile Include="Vector4Tests.cs" />
